Related published patent
﻿Patent abstract
Disclosed herein is a method of evaluating user knowledge using multiple-choice questions. Accordingly, the method may include retrieving, using a storage device, an assess- ment module and transmitting, using the communication device, the assessment module to a student device, wherein the assessment module comprises a multiple choice question and options. Further, the method may include receiving, using the communication device, bets corresponding to the options from the at least one student device, retrieving, using the storage device, an assessment answer module compris- ing correct probabilities, analyzing, using the processing device, the bets and the correct probabilities, generating, using the processing device, a knowledge score based on the analyzing of the plurality of bets and the correct probabili- ties, transmitting, using the communication device, the knowledge score to an examiner device and the student device, and storing, using the storage device, the bets and the knowledge score.